requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.12.1
openai==1.3.0
deepgram-sdk==3.2.7
python-magic==0.4.27
openai-whisper==20231117
numpy>=1.20.0
torch>=2.0.0